[refactor] canvas 성능 모니터 유틸 분리
